Added Parallel State Machine For Character Control
I implemented state machines for character movement and a few default states. The integration with the GDScript camera file was very easy. The prior movement script is still in the file structure unchanged so to revert to the previous movement system, it just has to be reattached to the player node.