Added interactable script
I added an interactable script and a scene for a moving platform with two buttons. To use the interactable script, you have to attach it to a node and set the id for the interaction you want to occur. Then on the same or different node, attach the script for the behavior your want and connect it to the interaction signal. Filter effects by the interaction id.