Updated save files
Save file now includes many new stats which will be used in the future (hoping that the format will not need to change past this). Also added the checkpoint system (which should have been a separate commit, but ended up doing both at the same time).

To add a checkpoint just add the checkpoint script to a node as well as marking it in groups as a checkpoint.