Changed playtime mechanism
No longer are the days in which man must track time himself; for technology has been invented. For technology; created at the hands of men, shall prevail.
Player no longer manages the playtime with deltatime adding. Now, the save file stores the time at which the file is loaded, and calculates the playtime when the file is saved.