Fixed issues with missing collisions due to layers
In the merge process, layers started to be used but were not added to flames and checkpoints, thus meaning they were never collided with and were never collected. This should be fixed now.

The fourth layer was added as "Collectables", which is mainly flames, but checkpoints are under that as well (at least for now) since they collide the same.