Changed the organization of flames
Now in data, flames are organized by their island. This allows us to get the total flames on one island from any scene related to the island. Useful if we want to display collected flames for each island on the island select map. Also useful if we need the total flames information in any scene that is not the main island (in the dungeon or perhaps subareas).
Documentation help is added to the Documentation document.