Slight adjustments to file deleting
Now, deleting a file will permanently delete it, rather than just moving it to the recycling bin. This can be changed if needed, the old code is just commented out now.

Also, added an emergency delete all files function. This is used when file select crashes due to a change in the save file JSON format.