Added FileUtility class
This class allows PlayerData and file_select to ignore certain details about loading and writing to files. This will also abstract the manner of interacting with files, which will be helpful for anything that needs to read or write files in the future, since the code would be duplicated anyways.