Fixed flames not correctly working with death.
The issue ended up being due to Godot not wanting to handle int to float comparisons. The flame id was being stored as an int, then when you write the save and read it again it was turned to a float. This caused issues if the value never went through the JSON process. Now, it will always be a float (to avoid weird JSON conversions.)