Added a draft of the first island and dungeon.
These ideas will need to be expanded upon, but I figured we need to start somewhere.

Basically, there are a number of things being showcased in this draft of the island.

The basic design of the tutorial area is present, leading into the dungeon.

The signs are temporary for loading zones, which I believe someone else is working on.

The boat is also present just to show how that might work.