New and improved Spike floor!!!
I remade the spike floor to use MultiMesh, as well as reworking it to be easier to work with. Now you just tell the spike floor what size for the floor and scale for the spikes. There is a rescale mode that fits the spikes to the floor size, and an overflow mode that ensures your scale size is retained. (if neither of them fit ur vibe, maybe try picking better values :D )