Added wind gust basic functionality
Added a basic idea of wind gusts to push the player in the direction of the wind.

However, I think a rework to the movement system is needed to make it actually work.

Right now, wind gusts need to affect velocity, but also walking resets velocity along with a few other things. Other parts of movement, like friction after moving, rely on not having velocity removed.