Added coyote time
No longer are the days in which walking off a platform kept all of your jumps.

Now, the player has a short time frame in which they can still use both jumps after starting to fall, but after that time frame, they lose one of the jumps (the normal one. you can still double jump in the air, but only one jump in the air. you get it.)

Currently set to 0.1 seconds, perhaps adjust slightly.