Tested out a potential mechanic for the lighthouse dungeon
Not sure if we want to do this, but it could be quite easy to add such a mechanic if needed.

The work here is done very quickly and loosely, but if we decide to go with the idea, probably just make a script like "mover.gd" as "rotator.gd" for reusable mechanics.

Objects connected to the outer wall of the lighthouse rotate, those unconnected do not.

Perhaps make a more visual indicator of what will rotate? (This could also enable rotating things not on the wall?)