Added a flames index
The flames index handles the details on all the flames of each island, which allows us to access the names of all flames on an island from anywhere, as well as list out the collected flames from anywhere.

Potentially this could be improved with a tool in Godot editor that allows us to automatically update name of flames somehow?