Added order challenges.
Added the challenge of interacting with a set of pylons in the correct order to spawn a flame. This likely could be expanded to allow order puzzles in dungeons as well, not just spawning flames.

The pylons exist even after the flame is collected, but the flame cannot be collected again.