Added pre- and post-dungeon tags
Now, objects can be added to the scene depending on whether the player has completed the dungeon or not.

I think having the flame door disappear after clearing the dungeon makes sense.

Plus, allows the boat to spawn after the ruins dungeon is completed, so you can only leave after beating the level.