Added a loading screen scene for the screen
When you die, or load into a new scene, there is a loading screen that fades in and out.

It will load the next scene in the background, and fade back in when ready.

Little details, the "loading..." and little icon only appear if the load is taking noticeable time.